On Linux, find the Playdate serial device if possible
This will use the /dev/serial/by-id/usb-Panic_Inc_Playdate_PDU1-*
symlink to find the appropriate serial device, if possible.  If the
PLAYDATE_SERIAL_DEVICE variable is set, that will take precedence.  If
multiple Playdate devices are found, the user is told, and the first is
used.  If none are found, the same /dev/ttyACM0 backup will be used.